Add documentation on async actions in aliases to README
After having the same problem in issue #70, I thought it might be helpful to have a complete solution documented in one place. Could also go into the Wiki on the advanced page on aliases, but figured this was a common enough use case to warrant placement in the basic readme file.