remove external messaging (connecting to other extensions)
Previously, we used ports to know which external
contexts were interested in store updates. We
don't use ports anymore, so there isn't an obvious
mechanism to know which external contexts are
interested in store updates.

This means webext-redux can't be used across
extensions anymore. I'm not sure how often this
was used [1], but if there's use-cases for this,
we could consider tracking contexts that recently
requested messages and broadcast state updates to
them, in addition to broadcasting state updates to
the open tabs and our own extension pages.

1: Tried searching GitHub but found no results.
This indicates the `extensionId` option is never
used in open-source code. Or I didn't use the
right search query.

https://github.com/search?q=webext-redux+extensionId&type=code